backend/requirements/prod.txt
django==4.2.11
environs==11.0.0
gunicorn==21.2.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-filter==24.2
Pillow==10.2.0
drf-yasg==1.21.7
django-cors-headers
setuptools
bcrypt==4.1.2
